BENTO-15. kiji script calls upgrade check-in helper scripts.
This commit modifies the kiji script so that it runs two scripts related
to checking in with the upgrade server, if we are running in a Kiji BentoBox.
The scripts are:

1. record-usage-time.sh, which should record the timestamps at which the kiji script
is run.
2. upgrade-informer.sh, which should tell the user that a new version of BentoBox is
available (on some periodic basis).

The scripts should be in the same directory as the kiji script if they are to be run.
This will be the case in BentoBox.

Signed-off-by: Aaron Kimball <aaron@wibidata.com>
commit: https://github.com/kijiproject/kiji-schema/commit/8a2e456c881603d237c4875ec87165b1cd0ab2ab