CHOP-45. Ensure library jars are shipped with the dcache of run jobs.
This commit introduces a tool that expands a classpath into a comma-separated
list of paths of jars to be shipped on the distributed cache of run jobs.
The chop script is updated to use this tool to set the value of the configuration
variable -Dtmpjars.

Reviewed by: robertchu
commit: https://github.com/kijiproject/kiji-express/commit/51b2e13d6d69f6b0774121c08af696dc9e89774a