CHOP-60. Fix the script tests so that they run in hadoop local mode.
Add a method to the ScriptRunnerSuite that modifies the classpath of the System
classloader. The reason for this horrible hack is because the classes produced
by the Eval package when compiling the script lives in a separate classloader
and hence when the M/R job is launched in local mode, Cascading can't find these
classes unless the folder containing the compiled classes is added to the
classpath.

Reviewed by: robert
commit: https://github.com/kijiproject/kiji-express/commit/72571ff175fda642ab9d15a1c47c3655188bd703