EXP-90. Proper configuration for jobs launched through schema-shell.
This commit modifies the execution of the model lifecycle through the
modeling schema shell model so that express library jars are shipped
with launched jobs. This is accomplished by setting a JVM property
containing a list of library jars through the express script, and then
reading that property when launching a job. Additionally,
configuration properties are set to disable schema validation for the
time being.

This commit also modifies things so that users can specify additional
library jars to ship with jobs not as URIs, but with absolute or
relative paths to the local file system.

Reviewed by: robertchu
commit: https://github.com/kijiproject/kiji-express/commit/5cf0bf1fb7178b7f68399cc73adcf236604f408f