SCHEMA-335. Enable nullable 'salt' field in RowKeyFormat2.
Removing the possibility for this field to be null is a regression;
If the user selects RowKeyEncoding.RAW, this field is ignored and may
be set to null. This is explicitly done by the Kiji shell.
commit: https://github.com/kijiproject/kiji-schema/commit/34de40d97df54edd821eeceea4ddea863c8bf79f