KIJIMR-242. Fix NPE on region servers with missing timestamp.
Ensure that when a new HFile is written that the timerange is written so that
HBase will properly bulk load and perform scans. Previously this was done only
when the locality group writer was closed and not when the writer was closed
when rolling a new HFile.

Also to support external packages in writing HFiles, made some previously
private static methods public.

Reviewer: taton
commit: https://github.com/kijiproject/kiji-mapreduce/commit/38026007958f5d133114264e803778c3c24714cd