KijiTableContext extends KijiDeleter.
There are two concrete implementations, which use different strategies
to satisfy the new interface.

DirectKijiTableWriterContext is the first, which trades its KijiPutter
in for a KijiBufferedWriter.  Once in place it delegates all of the
deleter signatures.

HFileWriterContext is the more meaty change.  Its implementation
proceeds in a few steps.

First, HFileKeyValue is augmented to allow it to represent deletes.
This is done by adding an inner enum named Type, which parallels
HBase's own KeyValue.Type.  Along with puts, it can represent cell,
column and (hbase) family deletes.  The appropriate type codes are
passed along to the underlying HBase KeyValues.

Next, HFileWriterContext implements most of its new signatures in the
obvious way: create an HBaseKeyValue of the appropriate type and write
it to underlying storage.

The deleteRow signatures are an exception to that pattern, since HBase
does not represent row deletions with individual (cross-family)
tombstones.  Instead, it is required to handle each locality group
individually.

Following similar functionality in HBaseKijiBufferedWriter, we walk
the table's layout and visit each locality group.  Where a locality
group contains a single (kiji) family, we issue a single (hbase)
family delete and are done.

When the locality group contains multiple families, we must consider
the type of the deleted family.  For group-type families we issue a
column delete for each column.

For map-type families, we are forced to query all qualifiers in the
row and issue column deletes for any we find.  Volume aside, this is
undesirable because it is susceptible to a race: cells created after
the HFiles are generated but before they are bulk loaded will still be
visible, contrary to user expectations.  Short of rejecting the
operation outright or the introduction of an HBase tombstone for
qualifier prefixes, we can do no better."

review: https://review.kiji.org/r/2166
contributed-by: phs (Philip Smith, phil.h.smith@gmail.com)
Signed-off-by: Lee Sheng <lsheng@wibidata.com>
commit: https://github.com/kijiproject/kiji-mapreduce/commit/60dcc5eb0531ef1837e6654ed97afd04cb6528d5